+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v3.3.0-20201021
 * StackStorm 3.3.0 released
 * Upgrade Ubuntu to `16.04.07` (#50)
+* Switch the VirtualBox guest additions installation from .iso to .deb packages (#49)
 
 ## v3.2.0-20200429
 * StackStorm 3.2.0 released